move the single undo checker function
