+++ editor/src/components/canvas/canvas-strategies/strategies/absolute-resize-bounding-box-strategy.spec.browser2.tsx
@@ -35,11 +35,7 @@ import {
   BakedInStoryboardVariableName,
   BakedInStoryboardUID,
 } from '../../../../core/model/scene-utils'
-import {
-  expectSingleUndoStep,
-  mouseClickAtPoint,
-  mouseDragFromPointWithDelta,
-} from '../../event-helpers.test-utils'
+import { mouseClickAtPoint, mouseDragFromPointWithDelta } from '../../event-helpers.test-utils'
 import { CanvasControlsContainerID } from '../../controls/new-canvas-controls'
 import { setFeatureEnabled } from '../../../../utils/feature-switches'
 import { CSSProperties } from 'react'
@@ -292,9 +288,7 @@ async function doDblClickTest(editor: EditorRenderResult, testId: string): Promi
 
   const nineBlockControlSegment = editor.renderedDOM.getByTestId(testId)
 
-  await expectSingleUndoStep(editor, async () =>
-    mouseClickAtPoint(nineBlockControlSegment, { x: 2, y: 30 }, { eventOptions: { detail: 2 } }),
-  )
+  mouseClickAtPoint(nineBlockControlSegment, { x: 2, y: 30 }, { eventOptions: { detail: 2 } })
 
   return div
 }
