+++ editor/src/components/inspector/add-remove-layout-system-control.spec.browser2.tsx
@@ -1,11 +1,8 @@
 import { setFeatureEnabled } from '../../utils/feature-switches'
 import { shiftModifier } from '../../utils/modifiers'
+import { expectSingleUndoStep } from '../../utils/utils.test-utils'
 import { CanvasControlsContainerID } from '../canvas/controls/new-canvas-controls'
-import {
-  expectSingleUndoStep,
-  mouseClickAtPoint,
-  pressKey,
-} from '../canvas/event-helpers.test-utils'
+import { mouseClickAtPoint, pressKey } from '../canvas/event-helpers.test-utils'
 import { renderTestEditorWithCode, EditorRenderResult } from '../canvas/ui-jsx.test-utils'
 import { AddRemoveLayouSystemControlTestId } from './add-remove-layout-system-control'
