+++ editor/src/components/inspector/inspector-strategies/remove-flex-convert-to-absolute-strategy.spec.browser2.tsx
@@ -1,6 +1,7 @@
 import { setFeatureEnabled } from '../../../utils/feature-switches'
+import { expectSingleUndoStep } from '../../../utils/utils.test-utils'
 import { CanvasControlsContainerID } from '../../canvas/controls/new-canvas-controls'
-import { expectSingleUndoStep, mouseClickAtPoint } from '../../canvas/event-helpers.test-utils'
+import { mouseClickAtPoint } from '../../canvas/event-helpers.test-utils'
 import { renderTestEditorWithCode, EditorRenderResult } from '../../canvas/ui-jsx.test-utils'
 import { AddRemoveLayouSystemControlTestId } from '../add-remove-layout-system-control'
